rework stress tests
