+++ ydb/tests/stress/olap_workload/workload/ya.make
@@ -1,0 +1,13 @@
+PY3_LIBRARY()
+
+PY_SRCS(
+    __init__.py
+)
+
+PEERDIR(
+    ydb/public/sdk/python
+    ydb/public/sdk/python/enable_v3_new_behavior
+    ydb/tests/stress/common
+)
+
+END()
