+++ ydb/tests/stress/olap_workload/ya.make
@@ -6,9 +6,7 @@ PY_SRCS(
 
 PEERDIR(
     ydb/tests/stress/common
-    ydb/public/sdk/python
-    ydb/public/sdk/python/enable_v3_new_behavior
-    library/python/monlib
+    ydb/tests/stress/olap_workload/workload
 )
 
 END()
