+++ ydb/tests/stress/oltp_workload/ya.make
@@ -5,10 +5,8 @@ PY_SRCS(
 )
 
 PEERDIR(
-    ydb/tests/stress/common    
-    ydb/public/sdk/python
-    ydb/public/sdk/python/enable_v3_new_behavior
-    library/python/monlib
+    ydb/tests/stress/common
+    ydb/tests/stress/oltp_workload/workload
 )
 
 END()
