+++ ydb/tests/stress/simple_queue/workload/ya.make
@@ -1,0 +1,12 @@
+PY3_LIBRARY()
+
+PY_SRCS(
+    __init__.py
+)
+
+PEERDIR(
+    library/python/monlib
+)
+
+
+END()
